LPS-53578 - update language package
