+++ modules/apps/site-navigation/site-navigation-language-web/src/com/liferay/site/navigation/language/web/configuration/LanguageWebConfigurationUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.language.web.configuration;
+package com.liferay.site.navigation.language.web.configuration;
 
 import com.liferay.portal.kernel.configuration.Configuration;
 import com.liferay.portal.kernel.configuration.ConfigurationFactoryUtil;
