+++ modules/apps/site-navigation/site-navigation-language-web/src/com/liferay/site/navigation/language/web/portlet/template/dependencies/portlet_display_template_icon_menu.ftl
