+++ modules/apps/site-navigation/site-navigation-language-web/src/com/liferay/site/navigation/language/web/upgrade/LanguageWebUpgrade.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.language.web.upgrade;
+package com.liferay.site.navigation.language.web.upgrade;
 
-import com.liferay.language.web.upgrade.v1_0_0.UpgradePortletId;
-import com.liferay.language.web.upgrade.v1_0_0.UpgradePortletPreferences;
+import com.liferay.site.navigation.language.web.upgrade.v1_0_0.UpgradePortletId;
+import com.liferay.site.navigation.language.web.upgrade.v1_0_0.UpgradePortletPreferences;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.service.ReleaseLocalService;
@@ -58,7 +58,7 @@ public class LanguageWebUpgrade {
 		upgradeProcesses.add(new UpgradePortletPreferences());
 
 		_releaseLocalService.updateRelease(
-			"com.liferay.language.web", upgradeProcesses, 1, 1, false);
+			"com.liferay.site.navigation.language.web", upgradeProcesses, 1, 1, false);
 	}
 
 	private ReleaseLocalService _releaseLocalService;
